Move global accordion controls into separate component
