+++ packages/11ty/_includes/components/page-header.js
@@ -8,16 +8,15 @@ const path = require('path')
  */
 module.exports = function(eleventyConfig) {
   const contributors = eleventyConfig.getFilter('contributors')
+  const globalAccordionControls = eleventyConfig.getFilter('globalAccordionControls')
   const pageTitle = eleventyConfig.getFilter('pageTitle')
   const slugify = eleventyConfig.getFilter('slugify')
 
-  const { accordion: accordionConfig } = eleventyConfig.globalData.config
   const { labelDivider } = eleventyConfig.globalData.config.pageTitle
   const { imageDir } = eleventyConfig.globalData.config.figures
 
   return function (params) {
     const {
-      accordion: accordionPageConfig={},
       byline_format: bylineFormat,
       image,
       label,
@@ -25,11 +24,6 @@ module.exports = function(eleventyConfig) {
       subtitle,
       title
     } = params
-
-    /**
-     * Merge page and global config for accordion
-     */
-    const accordionControls = Object.assign({...accordionConfig.globalControls}, {...accordionPageConfig.globalControls|| {}})
 
     const classes = ['quire-page__header', 'hero']
 
@@ -59,14 +53,6 @@ module.exports = function(eleventyConfig) {
         `
       : ''
 
-    const globalAccordionControls = accordionControls.enable
-      ? html`
-        <div class="global-accordion-controls">
-          <button class="global-accordion-expand-all">${accordionControls.expand}</button>
-          <button class="global-accordion-collapse-all">${accordionControls.collapse}</button>
-        </div>`
-      : ''
-
     return html`
       <section class="${classes}">
         <div class="hero-body">
@@ -76,7 +62,7 @@ module.exports = function(eleventyConfig) {
           </h1>
           ${contributorsElement}
         </div>
-        ${globalAccordionControls}
+        ${globalAccordionControls(params)}
       </section>
       ${imageElement}
     `
